# assets/docs/rider/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "### 说明\n接口采用非标准Restful API，所有http返回代码均为`200`，当返回为非`200`时应为network错误，需要及时排查。\n\u003cbr\u003e\n接口返回说明查看 **[返回](#返回)**\n\u003cbr\u003e\n图片/附件前缀 `https://cdn.auroraride.com/`\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 接口编号\n\n第一位代表接口端分类\n\n- M 管理端\n- R 骑手端\n- E 门店端\n- C 通用\n\n第二位代表子分类（36进制）\n\n后三位代表子编号（10进制）\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n\n### 认证\n项目接口使用简单认证，认证方式为`header`中添加对应的认证`token`\n|  header   |  类型  |  接口  |\n| :-----: | :----: | :--: |\n|  X-Rider-Token   |  string   |  骑手API  |\n| X-Manager-Token | string |  后台API  |\n|  X-Employee-Token   | string |  员工API  |\n\n\u003cbr /\u003e\n\n\u003cbr /\u003e\n\n### 返回\n\n一个标准的返回应包含以下结构\n\n|  字段   |  类型  |  必填  |  说明  |\n| :-----: | :----: | :--: | :--: |\n|  code   |  int   |  是  |  返回代码  |\n| message | string |  是  |  返回消息  |\n|  data   | object |  是  |  返回数据  |\n\n`code`代码取值说明\n\n| 十进制 | 十六进制 | 说明 |\n| :----: | :------: | :--: |\n| 0  |  0x000  | 请求成功 |\n| 256 |  0x100  | 请求失败 |\n| 512 |  0x200  | *需要认证(跳转登录) |\n| 768 |  0x300  | *用户被封禁 |\n| 1024 |  0x400  | 资源未获 |\n| 1280 |  0x500  | 未知错误 |\n| 1536 |  0x600  | *需要实名 |\n| 1792 |  0x700  | *需要验证 (更换设备, 需要人脸验证) |\n| 2048 |  0x800  | *需要联系人 |\n| 2304 |  0x900  | 请求过期 |\n\n\u003e 当返回值是`1792(0x700)需要人脸验证`或`1536(0x600)需要实名`的时候`data`返回值为`{\"url\": \"string\"}`, 直接跳转url\n\n\n比如：\n\u003e 默认成功返回\n```json\n{\n  \"code\": 0,\n  \"message\": \"OK\",\n  \"data\": {\n    \"status\": true\n  }\n}\n```\n",
...
    },
    "basePath": "/",
    "paths": {
...
        "/rider/v1/authenticator": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1003 实名认证",
                "operationId": "RiderAuthenticator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "desc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RiderContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.FaceAuthUrlResponse"
                        }
                    }
                }
            }
        },
        "/rider/v1/authenticator/{token}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1004 实名认证结果",
                "operationId": "RiderAuthResult",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "实名认证token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/contact": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1002 添加紧急联系人",
                "operationId": "RiderContact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "desc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RiderContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/deposit": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1007 获取已缴押金",
                "operationId": "RiderRiderDeposit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.RiderDepositRes"
                        }
                    }
                }
            }
        },
        "/rider/v1/deregister": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1008 注销账户",
                "operationId": "RiderRiderDeregister",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/face/{token}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1005 获取人脸校验结果",
                "operationId": "RiderFaceResult",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "人脸校验token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/profile": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1006 获取个人信息",
                "operationId": "RiderRiderProfile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.RiderSigninRes"
                        }
                    }
                }
            }
        },
        "/rider/v1/reserve": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reserve - 预约"
                ],
                "summary": "R8001 获取未完成预约",
                "operationId": "RiderReserveUnfinished",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功, 预约不存在时为`null`",
                        "schema": {
                            "$ref": "#/definitions/model.ReserveUnfinishedRes"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reserve - 预约"
                ],
                "summary": "R8002 创建预约",
                "operationId": "RiderReserveCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "预约信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReserveCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.ReserveUnfinishedRes"
                        }
                    }
                }
            }
        },
        "/rider/v1/reserve/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reserve - 预约"
                ],
                "summary": "R8003 取消预约",
                "operationId": "ManagerReserveCancel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "骑手校验token",
                        "name": "X-Rider-Token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "预约ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    }
                }
            }
        },
...
        "/rider/v1/signin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rider - 骑手"
                ],
                "summary": "R1001 登录或注册",
                "operationId": "RiderSignin",
                "parameters": [
                    {
                        "description": "desc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RiderSignupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/model.RiderSigninRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.City": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "城市ID",
                    "type": "integer"
                },
                "name": {
                    "description": "城市",
                    "type": "string"
                }
            }
        },
...
        "model.Enterprise": {
            "type": "object",
            "properties": {
                "agent": {
                    "description": "是否代理商模式",
                    "type": "boolean"
                },
                "days": {
                    "description": "代理商时间选项",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "enterpriseContactName": {
                    "description": "团签联系人",
                    "type": "string"
                },
                "enterpriseContactPhone": {
                    "description": "团签联系电话",
                    "type": "string"
                },
                "enterpriseName": {
                    "description": "团签名称",
                    "type": "string"
                },
                "id": {
                    "description": "企业ID",
                    "type": "integer"
                },
                "name": {
                    "description": "企业名称",
                    "type": "string"
                }
            }
        },
        "model.EnterpriseContact": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "联系姓名",
                    "type": "string"
                },
                "phone": {
                    "description": "联系电话",
                    "type": "string"
                }
            }
        },
...
        "model.EnterpriseStation": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "站点ID",
                    "type": "integer"
                },
                "name": {
                    "description": "站点名称",
                    "type": "string"
                }
            }
        },
...
        "model.FaceAuthUrlResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "description": "实名认证跳转链接",
                    "type": "string"
                }
            }
        },
...
        "model.ReserveCreateReq": {
            "type": "object",
            "properties": {
                "business": {
                    "description": "业务选项 active:激活, pause:寄存, continue:取消寄存, unsubscribe:退租",
                    "type": "string",
                    "enum": [
                        "active",
                        "pause",
                        "continue",
                        "unsubscribe"
                    ]
                },
                "cabinetId": {
                    "description": "电柜ID",
                    "type": "integer"
                }
            }
        },
        "model.ReserveStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6
            ],
            "x-enum-comments": {
                "ReserveStatusCancel": "已取消",
                "ReserveStatusFail": "已失败",
                "ReserveStatusInvalid": "已失效",
                "ReserveStatusPending": "已预约",
                "ReserveStatusProcessing": "进行中",
                "ReserveStatusSuccess": "已完成",
                "ReserveStatusTimeout": "已超时"
            },
            "x-enum-varnames": [
                "ReserveStatusPending",
                "ReserveStatusProcessing",
                "ReserveStatusSuccess",
                "ReserveStatusFail",
                "ReserveStatusTimeout",
                "ReserveStatusCancel",
                "ReserveStatusInvalid"
            ]
        },
        "model.ReserveUnfinishedRes": {
            "type": "object",
            "properties": {
                "business": {
                    "description": "预约业务 active:激活, pause:寄存, continue:取消寄存, unsubscribe:退租",
                    "type": "string"
                },
                "cabinetId": {
                    "description": "电柜ID",
                    "type": "integer"
                },
                "fid": {
                    "description": "设施ID",
                    "type": "string"
                },
                "id": {
                    "description": "预约ID",
                    "type": "integer"
                },
                "status": {
                    "description": "状态 0:已预约 1:进行中",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ReserveStatus"
                        }
                    ]
                },
                "time": {
                    "description": "预约时间",
                    "type": "string"
                }
            }
        },
...
        "model.RiderContact": {
            "type": "object",
            "required": [
                "name",
                "phone",
                "relation"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "relation": {
                    "type": "string"
                }
            }
        },
        "model.RiderDepositRes": {
            "type": "object",
            "properties": {
                "deposit": {
                    "type": "number"
                }
            }
        },
...
        "model.RiderSigninRes": {
            "type": "object",
            "properties": {
                "cabinetBusiness": {
                    "description": "是否可以自主使用电柜办理业务",
                    "type": "boolean"
                },
                "contact": {
                    "description": "联系人",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.RiderContact"
                        }
                    ]
                },
                "deposit": {
                    "description": "需缴押金",
                    "type": "number"
                },
                "enterprise": {
                    "description": "所属企业",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Enterprise"
                        }
                    ]
                },
                "enterpriseContact": {
                    "description": "团签联系方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.EnterpriseContact"
                        }
                    ]
                },
                "exitEnterprise": {
                    "description": "判断能否退出团签",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "isAuthed": {
                    "description": "是否已认证",
                    "type": "boolean"
                },
                "isContactFilled": {
                    "description": "联系人是否添加",
                    "type": "boolean"
                },
                "isNewDevice": {
                    "description": "是否新设备",
                    "type": "boolean"
                },
                "name": {
                    "description": "姓名, 实名认证后才会有",
                    "type": "string"
                },
                "orderNotActived": {
                    "description": "是否存在未激活订单",
                    "type": "boolean"
                },
                "phone": {
                    "description": "电话",
                    "type": "string"
                },
                "qrcode": {
                    "description": "二维码",
                    "type": "string"
                },
                "station": {
                    "description": "站点",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.EnterpriseStation"
                        }
                    ]
                },
                "subscribe": {
                    "description": "骑士卡",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Subscribe"
                        }
                    ]
                },
                "token": {
                    "description": "认证token",
                    "type": "string"
                },
                "useStore": {
                    "description": "是否可使用门店办理业务",
                    "type": "boolean"
                }
            }
        },
        "model.RiderSignupReq": {
            "type": "object",
            "required": [
                "phone",
                "smsCode",
                "smsId"
            ],
            "properties": {
                "cityId": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "smsCode": {
                    "type": "string"
                },
                "smsId": {
                    "type": "string"
                }
            }
        },
...
        "model.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "model.Subscribe": {
            "type": "object",
            "properties": {
                "alterDays": {
                    "description": "改动天数",
                    "type": "integer"
                },
                "business": {
                    "description": "是否可办理业务",
                    "type": "boolean"
                },
                "city": {
                    "description": "所属城市",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.City"
                        }
                    ]
                },
                "days": {
                    "description": "总天数 = 骑士卡天数 + 改动天数 + 暂停天数 + 续费天数 + 已缴纳逾期滞纳金天数",
                    "type": "integer"
                },
                "ebike": {
                    "description": "电车属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Ebike"
                        }
                    ]
                },
                "endAt": {
                    "description": "结束时间 / 预计套餐结束时间",
                    "type": "string"
                },
                "enterprise": {
                    "description": "企业信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Enterprise"
                        }
                    ]
                },
                "id": {
                    "description": "订阅ID",
                    "type": "integer"
                },
                "initialDays": {
                    "description": "初始购买骑士卡天数",
                    "type": "integer"
                },
                "intelligent": {
                    "description": "是否智能柜订阅",
                    "type": "boolean"
                },
                "model": {
                    "description": "可用电池型号",
                    "type": "string"
                },
                "models": {
                    "description": "可用电池型号",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BatteryModel"
                    }
                },
                "needContract": {
                    "description": "是否需要签约",
                    "type": "boolean"
                },
                "order": {
                    "description": "订单信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SubscribeOrderInfo"
                        }
                    ]
                },
                "overdueDays": {
                    "description": "已缴纳逾期滞纳金天数",
                    "type": "integer"
                },
                "pauseDays": {
                    "description": "暂停天数",
                    "type": "integer"
                },
                "plan": {
                    "description": "骑士卡信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Plan"
                        }
                    ]
                },
                "remaining": {
                    "description": "剩余天数 = 总天数 - 已过时间",
                    "type": "integer"
                },
                "riderId": {
                    "description": "骑手ID",
                    "type": "integer"
                },
                "startAt": {
                    "description": "开始时间",
                    "type": "string"
                },
                "status": {
                    "description": "状态 0未激活 1计费中 2寄存中 3已逾期 4已退订 5已取消",
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3,
                        4,
                        5
                    ]
                },
                "suspend": {
                    "description": "是否暂停扣费",
                    "type": "boolean"
                }
            }
        },
...
    }
}